Move the React to its own module
This will make it easier to introduce multiple Reactors, such as a no-op
one (to use glommio for its task scheduling, TPC capabilities) or one
suited for a different platform, like macOS.
