+++ glommio/src/net/stream.rs
@@ -4,7 +4,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2020 Datadog, Inc.
 //
 use crate::{
-    parking::Reactor,
+    reactor::Reactor,
     sys::{self, DmaBuffer, Source, SourceType},
     ByteSliceMutExt,
     Local,
